Make REG_MDATA a single byte

+++ src/platform_ctr.c
@@ -21,7 +21,7 @@
 #include "../include/ncgc/ntrcard.h"
 
 #define REG_MCNT                (*(volatile uint16_t *)0x10164000)
-#define REG_MDATA               (*(volatile uint16_t *)0x10164002)
+#define REG_MDATA               (*(volatile uint8_t *)0x10164002)
 #define REG_ROMCNT              (*(volatile uint32_t *)0x10164004)
 #define REG_CMDP                ((volatile uint8_t *)0x10164008)
 #define REG_CMD                 (*(volatile uint64_t *)0x10164008)
